--- right-for-life_front/src/app/App.jsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import AnimalsPage from './containers/AnimalsPage/AnimalsPage';
-
-function App() {
-  return (
-    <AnimalsPage />
-  );
-}
-
-export default App;
