+++ right-for-life_front/src/components/AnimalCard/index.js
@@ -1,0 +1,1 @@
+export { AnimalCard } from './AnimalCard';
