+++ right-for-life_front/src/components/AnimalsList/index.js
@@ -1,0 +1,1 @@
+export { AnimalsList } from './AnimalsList';
