+++ right-for-life_front/src/components/Footer/index.js
@@ -1,0 +1,1 @@
+export { Footer } from './Footer';
