+++ right-for-life_front/src/containers/AnimalsPage/index.js
@@ -1,0 +1,1 @@
+export { AnimalsPage } from './AnimalsPage';
